requirements.txt
asgiref==3.4.1
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
Faker==15.1.1
python-dateutil==2.8.2
pytz==2022.1
sqlparse==0.4.2
